Fixed comment (thx to Unrated)
git-svn-id: svn://gallium.openminds.be/opt/svn/fork2/trunk@2268 869247f0-a1f3-4d5c-9449-4b73fc252fa6

+++ default_www/backend/modules/content_blocks/actions/edit.php
@@ -172,7 +172,7 @@ class BackendContentBlocksEdit extends BackendBaseActionEdit
 				$item['hidden'] = $this->frm->getField('hidden')->getValue();
 
 				// insert the item
-				$id = (int) BackendContentBlocksModel::update($this->id, $item);
+				$id = BackendContentBlocksModel::update($this->id, $item);
 
 				// everything is saved, so redirect to the overview
 				$this->redirect(BackendModel::createURLForAction('index') .'&report=edited&var='. urlencode($item['title']) .'&highlight=id-'. $id);
